+++ src/main/java/com/vet/vet/BackEnd/business/concretes/DoctorManager.java
@@ -98,11 +98,15 @@ public class DoctorManager implements IDoctorService {
         // If the result variable returns true, this means that no error occurred during the deletion process.
         Boolean result = false;
 
-        try {
-            this.doctorRepository.deleteById(id);
-            result = true;
-        }catch (Exception e){
-            System.out.println(e.getMessage());
+        if (this.doctorRepository.existsById(id)){
+            try {
+                this.doctorRepository.deleteById(id);
+                result = true;
+            }catch (Exception e){
+                System.out.println(e.getMessage());
+            }
+        }else{
+            throw  new RuntimeException(("there is no record with this id in database : " + id));
         }
 
         return result;
